+++ hockey_scraper/__init__.py
@@ -1,6 +1,7 @@
-#from .nwhl import scrape_functions as nwhl
 from .nhl.live_scrape import ScrapeLiveGames, LiveGame
 from .nhl.scrape_functions import scrape_games, scrape_date_range, scrape_seasons, scrape_schedule
 from .nhl import live_scrape
 from .utils import shared
 from . import utils
+
+#from .nwhl import scrape_schedule as nwhl_scrape_schedule