+++ hockey_scraper/utils/save_pages.py
@@ -17,9 +17,9 @@ def create_base_file_path(file_info):
     :return: path 
     """
     # Shitty fix for when you already have it saved but don't have nwhl folders
-    # if 'nwhl' in file_info['type']:
-    #     if not os.path.isdir(os.path.join(file_info['dir'], 'docs', str(file_info['season']), file_info['type'])):
-    #         os.mkdir(os.path.join(file_info['dir'], 'docs', str(file_info['season']), file_info['type']))
+    if 'nwhl' in file_info['type']:
+        if not os.path.isdir(os.path.join(file_info['dir'], 'docs', str(file_info['season']), file_info['type'])):
+            os.mkdir(os.path.join(file_info['dir'], 'docs', str(file_info['season']), file_info['type']))
 
     return os.path.join(file_info['dir'], 'docs', str(file_info['season']), file_info['type'], file_info['name'] + ".txt")